+++ src/actions/wallet/deployContract.ts
@@ -1,9 +1,10 @@
 import type { Abi } from 'abitype'
 
 import type { Account } from '../../accounts/types.js'
-import type { Chain, GetChain } from '../../chains/types.js'
+import type { Chain } from '../../chains/types.js'
 import type { Client } from '../../clients/createClient.js'
 import type { Transport } from '../../clients/transports/createTransport.js'
+import type { GetChain } from '../../types/chain.js'
 import type { GetConstructorArgs } from '../../types/contract.js'
 import type { Hex } from '../../types/misc.js'
 import type { UnionOmit } from '../../types/utils.js'
