+++ src/actions/wallet/writeContract.ts
@@ -1,10 +1,11 @@
 import type { Abi } from 'abitype'
 
 import type { Account } from '../../accounts/types.js'
-import type { Chain, GetChain } from '../../chains/types.js'
+import type { Chain } from '../../chains/types.js'
 import type { Client } from '../../clients/createClient.js'
 import type { Transport } from '../../clients/transports/createTransport.js'
 import type { GetAccountParameter } from '../../types/account.js'
+import type { GetChain } from '../../types/chain.js'
 import type { ContractFunctionConfig, GetValue } from '../../types/contract.js'
 import type { Hex } from '../../types/misc.js'
 import type { UnionOmit } from '../../types/utils.js'
